+++ src/muibook/story/components/mui-dropdown/index.js
@@ -387,7 +387,7 @@ class storyDropdown extends HTMLElement {
                 Card Artwork
                 <mui-icon-add slot="after"></mui-icon-add>
             </mui-button>
-            <mui-menu>
+            <mui-menu width="max-content">
               <mui-v-stack space="var(--space-300)" style="padding: var(--space-300)">
               <mui-smart-card
               type="Debit"
@@ -410,7 +410,7 @@ class storyDropdown extends HTMLElement {
           &nbsp;&nbsp;&nbsp;&nbsp;Card Artwork<br>
           &nbsp;&nbsp;&nbsp;&nbsp;&lt;mui-icon-add slot="after"&gt;&lt;/mui-icon-add&gt;<br>
           &nbsp;&nbsp;&lt;/mui-button&gt;<br><br />
-          &nbsp;&nbsp;&lt;mui-menu&gt;<br />
+          &nbsp;&nbsp;&lt;mui-menu width=&quot;max-content&quot;&gt;<br />
           &nbsp;&nbsp;&lt;mui-v-stack space="var(--space-300)" style="padding: var(--space-300)"&gt;<br>
           &nbsp;&nbsp;&nbsp;&nbsp;&lt;mui-smart-card<br>
           &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;type="Debit"<br>
@@ -473,7 +473,7 @@ class storyDropdown extends HTMLElement {
                 <mui-icon-subtract slot="end"></mui-icon-subtract>
               </mui-icon-toggle>
             </mui-button>
-            <mui-menu>
+            <mui-menu width="max-content">
               <mui-v-stack space="var(--space-300)" style="padding: var(--space-300)">
               <mui-smart-card
               type="Debit"
@@ -500,7 +500,7 @@ class storyDropdown extends HTMLElement {
           &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;mui-icon-subtract slot="end"&gt;&lt;/mui-icon-subtract&gt;<br>
           &nbsp;&nbsp;&nbsp;&nbsp;&lt;/mui-icon-toggle&gt;<br>
           &nbsp;&nbsp;&lt;/mui-button&gt;<br><br />
-          &nbsp;&nbsp;&lt;mui-menu&gt;<br />
+          &nbsp;&nbsp;&lt;mui-menu width=&quot;max-content&quot;&gt;<br />
           &nbsp;&nbsp;&lt;mui-v-stack space="var(--space-300)" style="padding: var(--space-300)"&gt;<br>
           &nbsp;&nbsp;&nbsp;&nbsp;&lt;mui-smart-card<br>
           &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;type="Debit"<br>